+++ src/app/modules/auth/auth.module.ts
@@ -3,7 +3,7 @@ import { NgxCaptchaModule } from 'ngx-captcha';
 import { SharedModule } from 'src/app/shared/shared.module';
 import { AutRoutingModule } from './auth-routing.module';
 import { LoginNetworkComponent } from './components/login-network/login-network.component';
-import { ForgetPasswordPageComponent } from './pages/forget-password-page/forget-password-page.component';
+import { ForgotPasswordPageComponent } from './pages/forgot-password-page/forgot-password-page.component';
 import { SignInPageComponent } from './pages/sign-in-page/sign-in-page.component';
 import { SignUpPageComponent } from './pages/sign-up-page/sign-up-page.component';
 
@@ -16,7 +16,7 @@ import { SignUpPageComponent } from './pages/sign-up-page/sign-up-page.component
   declarations: [
     SignInPageComponent,
     SignUpPageComponent,
-    ForgetPasswordPageComponent,
+    ForgotPasswordPageComponent,
     LoginNetworkComponent,    
   ]
 })